+++ jiant/tasks/lib/mlm_pretokenized.py
@@ -1,23 +1,14 @@
-import numpy as np
 from dataclasses import dataclass
 from typing import List
 
 import jiant.utils.python.io as py_io
-from jiant.tasks.lib.templates import mlm as mlm_template
 from jiant.tasks.core import (
     Task,
     TaskTypes,
     BaseExample,
-    BaseTokenizedExample,
-    BaseDataRow,
-    BatchMixin,
 )
-from jiant.tasks.utils import ExclusiveSpan, truncate_sequences
-from jiant.tasks.lib.templates.shared import (
-    construct_single_input_tokens_and_segment_ids,
-    create_input_set_from_tokens_and_segments,
-    pad_single_with_feat_spec,
-)
+from jiant.tasks.utils import ExclusiveSpan
+import jiant.tasks.lib.templates.mlm_premasked as mlm_premasked_template
 
 
 @dataclass
@@ -57,68 +48,18 @@ class Example(BaseExample):
 
 
 @dataclass
-class TokenizedExample(BaseTokenizedExample):
-    guid: str
-    masked_tokens: List
-    label_tokens: List
-
-    def featurize(self, tokenizer, feat_spec):
-        # Handle masked_tokens
-        unpadded_masked_inputs = construct_single_input_tokens_and_segment_ids(
-            input_tokens=self.masked_tokens, tokenizer=tokenizer, feat_spec=feat_spec,
-        )
-        masked_input_set = create_input_set_from_tokens_and_segments(
-            unpadded_tokens=unpadded_masked_inputs.unpadded_tokens,
-            unpadded_segment_ids=unpadded_masked_inputs.unpadded_segment_ids,
-            tokenizer=tokenizer,
-            feat_spec=feat_spec,
-        )
-        # Handle label_tokens
-        special_tokens_count = 2  # CLS, SEP
-        pad_token = tokenizer.pad_token
-        unpadded_label_tokens, = truncate_sequences(
-            tokens_ls=[self.label_tokens],
-            max_length=feat_spec.max_seq_length - special_tokens_count,
-        )
-        if feat_spec.cls_token_at_end:
-            unpadded_label_tokens = unpadded_label_tokens + [pad_token, pad_token]
-        else:
-            unpadded_label_tokens = [pad_token] + unpadded_label_tokens + [pad_token]
-        unpadded_label_token_ids = tokenizer.convert_tokens_to_ids(unpadded_label_tokens)
-        masked_lm_labels = pad_single_with_feat_spec(
-            ls=unpadded_label_token_ids, feat_spec=feat_spec, pad_idx=feat_spec.pad_token_id,
-        )
-        masked_lm_labels = np.array(masked_lm_labels)
-        masked_lm_labels[masked_lm_labels == feat_spec.pad_token_id] = \
-            mlm_template.NON_MASKED_TOKEN_LABEL_ID
-        return DataRow(
-            guid=self.guid,
-            masked_input_ids=np.array(masked_input_set.input_ids),
-            input_mask=np.array(masked_input_set.input_mask),
-            segment_ids=np.array(masked_input_set.segment_ids),
-            masked_lm_labels=masked_lm_labels,
-            masked_tokens=unpadded_masked_inputs.unpadded_tokens,
-            label_tokens=unpadded_label_tokens,
-        )
+class TokenizedExample(mlm_premasked_template.TokenizedExample):
+    pass
 
 
 @dataclass
-class DataRow(BaseDataRow):
-    guid: str
-    masked_input_ids: np.ndarray
-    input_mask: np.ndarray
-    segment_ids: np.ndarray
-    masked_lm_labels: np.ndarray
-    masked_tokens: List
-    label_tokens: List
+class DataRow(mlm_premasked_template.BaseDataRow):
+    pass
 
 
 @dataclass
-class Batch(BatchMixin, mlm_template.BaseMLMBatch):
-    def get_masked(self, mlm_probability, tokenizer, do_mask):
-        assert mlm_probability is None
-        assert not do_mask
-        return self
+class Batch(mlm_premasked_template.Batch):
+    pass
 
 
 class MLMPretokenizedTask(Task):
@@ -148,6 +89,6 @@ class MLMPretokenizedTask(Task):
         for i, row in enumerate(py_io.read_jsonl(path)):
             yield Example(
                 guid="%s-%s" % (set_type, i),
-                tokenized_text=row["text"],
+                tokenized_text=row["tokenized_text"],
                 masked_spans=[ExclusiveSpan(start, end) for start, end in row["masked_spans"]],
             )
